version control working

+++ view/inc/version_control.php
@@ -16,7 +16,7 @@ if(file_exists(PATH_HOME . "_config/updates/version.txt")) {
     $actual = file_get_contents(PATH_HOME . "composer.lock");
     if($old !== $actual) {
         $conf = file_get_contents(PATH_HOME . "_config/config.php");
-        $version = (float) explode("')", explode("'VERSION', '", $conf)[1])[0];
+        $version = explode("')", explode("'VERSION', '", $conf)[1])[0];
         $newVersion = $version + 0.01;
         $conf = str_replace("'VERSION', '{$version}')", "'VERSION', '{$newVersion}')", $conf);
         $f = fopen(PATH_HOME . "_config/config.php", "w");
